refactor: Remove the possibility of nesting from API to simplify
Previously, it was able to nest using directives, which is no longer supported. Now, the change is reflected in the API
