--- src/main/java/com/virtuslab/using_directives/custom/utils/ast/SettingDefOrUsingValue.java
@@ -1,11 +1,0 @@
-package com.virtuslab.using_directives.custom.utils.ast;
-
-import com.virtuslab.using_directives.custom.utils.Position;
-
-public abstract class SettingDefOrUsingValue extends UsingTree {
-  public SettingDefOrUsingValue(Position position) {
-    super(position);
-  }
-
-  public SettingDefOrUsingValue() {}
-}
