+++ src/main/java/com/virtuslab/using_directives/custom/utils/ast/UsingValue.java
@@ -2,7 +2,7 @@ package com.virtuslab.using_directives.custom.utils.ast;
 
 import com.virtuslab.using_directives.custom.utils.Position;
 
-public abstract class UsingValue extends SettingDefOrUsingValue {
+public abstract class UsingValue extends UsingTree {
   public UsingValue(Position position) {
     super(position);
   }
