+++ src/test/resources/parser_tests/asts/ast14.json
@@ -1,16 +1,10 @@
 {
   "usingDefs": [
     {
-      "settingDefs": {
-        "settings": [
-          {
-            "key": "lib.cs",
-            "value": {
-              "type": "string",
-              "value": "org.scalameta :: munit : 0.4.3"
-            }
-          }
-        ]
+      "key": "lib.cs",
+      "value": {
+        "type": "string",
+        "value": "org.scalameta :: munit : 0.4.3"
       }
     }
   ],
