+++ src/test/resources/parser_tests/asts/ast4.json
@@ -1,17 +1,11 @@
 {
   "usingDefs": [
     {
-      "settingDefs": {
-        "settings": [
-          {
-            "key": "lib.cs",
-            "value": {
-              "type": "string",
-              "value": "org.scalameta :: munit : 0.4.3"
-            }
-          }
-        ]
-      }
+      "key": "lib.cs",
+      "value": {
+        "type": "string",
+        "value": "org.scalameta :: munit : 0.4.3"
+      }         
     }
   ],
   "codeOffset": 45,
