+++ src/test/resources/parser_tests/asts/ast5.json
@@ -1,49 +1,31 @@
 {
   "usingDefs": [
     {
-      "settingDefs": {
-        "settings": [
-          {
-            "key": "lib.cs",
-            "value": {
-              "type": "string",
-              "value": "org.scalameta :: munit : 0.4.3"
-            }
-          }
-        ]
+      "key": "lib.cs",
+      "value": {
+        "type": "string",
+        "value": "org.scalameta :: munit : 0.4.3"
       }
     },
     {
-      "settingDefs": {
-        "settings": [
-          {
-            "key": "lib.cs",
-            "value": {
-              "type": "string",
-              "value": "com.atlassian.commonmark:commonmark:0.13.1"
-            }
-          }
-        ]
+      "key": "lib.cs",
+      "value": {
+        "type": "string",
+        "value": "com.atlassian.commonmark:commonmark:0.13.1"
       }
     },
     {
-      "settingDefs": {
-        "settings": [
+      "key": "compiler.setting",
+      "value": {
+        "type": "list",
+        "value": [
           {
-            "key": "compiler.setting",
-            "value": {
-              "type": "list",
-              "value": [
-                {
-                  "type": "string",
-                  "value": "-Xfatal-warnings"
-                },
-                {
-                  "type": "string",
-                  "value": "-d ~/classes"
-                }
-              ]
-            }
+            "type": "string",
+            "value": "-Xfatal-warnings"
+          },
+          {
+            "type": "string",
+            "value": "-d ~/classes"
           }
         ]
       }
